Create a custom Hook to separate logic for managing the state for Application component
