+++ src/components/Application.js
@@ -1,62 +1,20 @@
-import React, { useState, useEffect } from "react";
-import axios from "axios";
+import React from "react";
 import DayList from "./DayList";
 import Appointment from "components/Appointment";
 import { getAppointmentsForDay, getInterviewersForDay, getInterview } from "helpers/selectors";
+import useApplicationData from "hooks/useApplicationData";
 import "components/Application.scss";
 
-export default function Application(props) {
-  const [state, setState] = useState({
-    day: "Monday",
-    days: [],
-    appointments: {},
-    interviewers: {}
-  });
+export default function Application() {
+  const {
+    state,
+    setDay,
+    bookInterview,
+    cancelInterview
+  } = useApplicationData();
 
-  const setDay = day => setState({ ...state, day });
-
-  const bookInterview = (id, interview) => {
-    // create a new object to copy specific (id) appointment and add/replace interview data
-    // interview data passed from Appointment component
-    const appointment = {
-      ...state.appointments[id],
-      interview: { ...interview }
-    };
-
-    // create a new object to copy all the appointments and add/replace appointment data
-    const appointments = {
-      ...state.appointments,
-      [id]: appointment
-    };
-
-    // 
-    return (
-      axios.put(`/api/appointments/${id}`, appointment)
-        .then((res) => setState((prev) => ({ ...prev, appointments })))
-    );
-  };
-
-  const cancelInterview = (id) => {
-    const appointment = {
-      ...state.appointments[id],
-      interview: null
-    };
-
-    const appointments = {
-      ...state.appointments,
-      [id]: appointment
-    }
-    
-    return (
-      axios.delete(`/api/appointments/${id}`)
-        .then((res) => setState((prev) => ({ ...prev, appointments })))
-    );
-  };
-
-  const dailyAppointments = getAppointmentsForDay(state, state.day);
   const dailyInterviewers = getInterviewersForDay(state, state.day);
-  // const dailyInterviewers = getInterviewersForDay(state, state.day);
-  // console.log(dailyInterviewers)
+  const dailyAppointments = getAppointmentsForDay(state, state.day);
 
   const parsedAppointments = dailyAppointments.map((appointment) => {
     const interview = getInterview(state, appointment.interview);
@@ -70,18 +28,6 @@ export default function Application(props) {
       cancelInterview={cancelInterview}
     />
   });
-
-  useEffect(() => {
-    Promise.all([
-      axios.get("/api/days"),
-      axios.get("/api/appointments"),
-      axios.get("/api/interviewers")
-    ])
-      .then((all) => {
-        setState(prev => ({ ...prev, days: all[0].data, appointments: all[1].data, interviewers: all[2].data }));
-      })
-      .catch((err) => console.log(err.message));
-  }, []);
 
   return (
     <main className="layout">
