+++ src/hooks/useVisualMode.js
@@ -22,7 +22,7 @@ function useVisualMode(initial) {
       setHistory(prev => [...prev.slice(0, prev.length - 1)]);
     }
   }
-  console.log(history)
+
   return {
     mode,
     transition,
